Added the "delete data" function and fixed the "add data" function
I added the “delete data” function, which asks for a table, the value
to remove and the variable that value is from. I also fixed the “add
data” function so that it will ask for the type of the input then use
that to alter the value input, in case it needs to add quotation marks.